Save files using threads

+++ src/main/java/com/example/app/photos/data/ImageResource.java
@@ -20,7 +20,7 @@ public class ImageResource extends BaseEntity {
   @Column
   private int fileSizesFlags;
   @Column
-  private int originalImageSize; //bytes
+  private long originalImageSize; //bytes
   @JsonIgnore
   @ManyToOne
   @JoinColumn(name="album_id", nullable=false)
